requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
pymongo==4.5.0
bcrypt==4.0.1
python-dotenv==1.0.0
Werkzeug==2.3.7
pip install flask-wtf
